% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROJ-package.R
\docType{package}
\name{projmin-package}
\alias{projmin}
\alias{projmin-package}
\title{projmin: Test Proj Minimally}
\description{
'PROJ' testing on travis
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com} (0000-0002-2471-7511)

}
\keyword{internal}
